+++ src/b42/lua/server/MoonshineOnDrinkActions.lua
@@ -13,7 +13,7 @@ local pzversion = string.sub(getCore():getVersionNumber(), 1, 2)
 
 -- PerkLevelup creates level up for perk.
 function PerkLevelup(character, perkType)
-    local perkLevel = character:getPerkLevel(perkType);
+    local perkLevel = character:getPerkLevel(perkType)
 
     if perkLevel < 10 then
         logger.Debug("levelup by moonshine drinking")
@@ -25,13 +25,13 @@ function PerkLevelup(character, perkType)
             xpInLevel = 0
         end
 
-        character:LevelPerk(perkType, false);
-        character:getXp():setXPToLevel(perkType, character:getPerkLevel(perkType));
+        character:LevelPerk(perkType, false)
+        character:getXp():setXPToLevel(perkType, character:getPerkLevel(perkType))
         SyncXp(character)
 
         if SandboxVars.Moonshine.KeepExperience then
             -- Add all XP: kicks with Type15 xp is large.
-            character:getXp():AddXPNoMultiplier(perkType, xpInLevel);
+            character:getXp():AddXPNoMultiplier(perkType, xpInLevel)
             SyncXp(character)
         end
     end
@@ -44,7 +44,7 @@ function OnDrink_HastyHerring(food, character, percent)
         return
     end
 
-    PerkLevelup(character, Perks.Sprinting);
+    PerkLevelup(character, Perks.Sprinting)
 end
 
 -- OnDrink_DoubleHastyHerring adds action to drink Moonshine.DoubleHastyHerring.
@@ -54,8 +54,8 @@ function OnDrink_DoubleHastyHerring(food, character, percent)
         return
     end
 
-    PerkLevelup(character, Perks.Sprinting);
-    PerkLevelup(character, Perks.Fitness);
+    PerkLevelup(character, Perks.Sprinting)
+    PerkLevelup(character, Perks.Fitness)
 end
 
 -- OnDrink_GreedyHammer adds action to drink Moonshine.GreedyHammer.
@@ -65,7 +65,7 @@ function OnDrink_GreedyHammer(food, character, percent)
         return
     end
 
-    PerkLevelup(character, Perks.Blunt);
+    PerkLevelup(character, Perks.Blunt)
 end
 
 -- OnDrink_DoubleGreedyHammer adds action to drink Moonshine.DoubleGreedyHammer.
@@ -75,8 +75,8 @@ function OnDrink_DoubleGreedyHammer(food, character, percent)
         return
     end
 
-    PerkLevelup(character, Perks.Blunt);
-    PerkLevelup(character, Perks.Strength);
+    PerkLevelup(character, Perks.Blunt)
+    PerkLevelup(character, Perks.Strength)
 end
 
 -- OnDrink_GreedyAxe adds action to drink Moonshine.GreedyAxe.
@@ -86,7 +86,7 @@ function OnDrink_GreedyAxe(food, character, percent)
         return
     end
 
-    PerkLevelup(character, Perks.Axe);
+    PerkLevelup(character, Perks.Axe)
 end
 
 -- OnDrink_DoubleGreedyAxe adds action to drink Moonshine.DoubleGreedyAxe.
@@ -96,8 +96,8 @@ function OnDrink_DoubleGreedyAxe(food, character, percent)
         return
     end
 
-    PerkLevelup(character, Perks.Axe);
-    PerkLevelup(character, Perks.Strength);
+    PerkLevelup(character, Perks.Axe)
+    PerkLevelup(character, Perks.Strength)
 end
 
 -- OnDrink_StrayBullet adds action to drink Moonshine.StrayBullet.
@@ -107,7 +107,7 @@ function OnDrink_StrayBullet(food, character, percent)
         return
     end
 
-    PerkLevelup(character, Perks.Aiming);
+    PerkLevelup(character, Perks.Aiming)
 end
 
 -- OnDrink_SlipperyFish adds action to drink Moonshine.SlipperyFish.
@@ -118,7 +118,7 @@ function OnDrink_SlipperyFish(food, character, percent)
         return
     end
 
-    PerkLevelup(character, Perks.Nimble);
+    PerkLevelup(character, Perks.Nimble)
 end
 
 -- OnDrink_SolidAdventurer adds action to drink Moonshine.SolidAdventurer.
@@ -129,7 +129,7 @@ function OnDrink_SolidAdventurer(food, character, percent)
         return
     end
 
-    PerkLevelup(character, Perks.Maintenance);
+    PerkLevelup(character, Perks.Maintenance)
 end
 
 -- OnDrink_SlenderDoe adds action to drink Moonshine.SlenderDoe.
@@ -140,22 +140,22 @@ function OnDrink_SlenderDoe(food, character, percent)
         return
     end
 
-    character:getNutrition():setWeight(SandboxVars.Moonshine.SlenderDoeSetWeight);
+    character:getNutrition():setWeight(SandboxVars.Moonshine.SlenderDoeSetWeight)
 
     if character:hasTrait(CharacterTrait.OVERWEIGHT) then
-        character:getCharacterTraits():remove(CharacterTrait.OVERWEIGHT);
+        character:getCharacterTraits():remove(CharacterTrait.OVERWEIGHT)
     end
 
     if character:hasTrait(CharacterTrait.UNDERWEIGHT) then
-        character:getCharacterTraits():remove(CharacterTrait.UNDERWEIGHT);
+        character:getCharacterTraits():remove(CharacterTrait.UNDERWEIGHT)
     end
 
     if character:hasTrait(CharacterTrait.OBESE) then
-        character:getCharacterTraits():remove(CharacterTrait.OBESE);
+        character:getCharacterTraits():remove(CharacterTrait.OBESE)
     end
 
     if character:hasTrait(CharacterTrait.VERY_UNDERWEIGHT) then
-        character:getCharacterTraits():remove(CharacterTrait.VERY_UNDERWEIGHT);
+        character:getCharacterTraits():remove(CharacterTrait.VERY_UNDERWEIGHT)
     end
 end
 
@@ -167,9 +167,8 @@ function OnDrink_NicotineOverdose(food, character, percent)
     end
 
     if character:hasTrait(CharacterTrait.SMOKER) then
-        character:getCharacterTraits():remove(CharacterTrait.SMOKER);
-        character:getStats():setStressFromCigarettes(0);
-        character:setTimeSinceLastSmoke(0);
+        character:getCharacterTraits():remove(CharacterTrait.SMOKER)
+        character:setTimeSinceLastSmoke(0)
     end
 end
 
@@ -180,16 +179,15 @@ function OnDrink_GreedySalvation(food, character, percent)
         return
     end
 
-    local bodyDamage = character:getBodyDamage();
+    local bodyDamage = character:getBodyDamage()
 
-    bodyDamage:setInfected(false);
-    bodyDamage:setInfectionMortalityDuration(-1);
-    bodyDamage:setInfectionTime(-1);
-    bodyDamage:setInfectionLevel(0);
+    bodyDamage:setInfected(false)
+    bodyDamage:setInfectionMortalityDuration(-1)
+    bodyDamage:setInfectionTime(-1)
 
-    local bodyParts = bodyDamage:getBodyParts();
+    local bodyParts = bodyDamage:getBodyParts()
     for i=bodyParts:size()-1, 0, -1  do
-        local bodyPart = bodyParts:get(i);
-        bodyPart:SetInfected(false);
+        local bodyPart = bodyParts:get(i)
+        bodyPart:SetInfected(false)
     end
 end