+++ src/b42/lua/server/MoonshineRecipesServer.lua
@@ -23,12 +23,14 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 time = 300,
                 cookingSkill = 3,
                 usedItems = {
-                    ["Base.WhiskeyWaterFull"] = 1,
                     ["Base.Yeast"] = 1,
                     ["Base.Corn"] = 1,
                     ["Base.UnusableWood"] = 1,
                 },
-                result = "Base.WhiskeyFull",
+                fluids = {
+                    {code = "Base.BottleCrafted", name = "Water", amount = 1},
+                },
+                result = "Base.Whiskey",
                 additionalResults = {},
             },
             {
@@ -40,9 +42,11 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 time = 300,
                 cookingSkill = 3,
                 usedItems = {
-                    ["Base.WineWaterFull"] = 1,
                     ["Base.Sugar"] = 1,
                     ["Base.Grapes"] = 1,
+                },
+                fluids = {
+                    {code = "Base.BottleCrafted", name = "Water", amount = 1},
                 },
                 result = "Base.Wine2",
                 additionalResults = {},
@@ -56,9 +60,11 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 time = 300,
                 cookingSkill = 3,
                 usedItems = {
-                    ["Base.BeerWaterFull"] = 1,
                     ["Base.Sugar"] = 1,
                     ["Base.Corn"] = 1,
+                },
+                fluids = {
+                    {code = "Base.BeerEmpty", name = "Water", amount = 1},
                 },
                 result = "Base.BeerBottle",
                 additionalResults = {},
@@ -75,11 +81,13 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 cookingSkill = 10,
                 usedItems = {
                     ["Moonshine.ExclusiveRecipe"] = 1,
-                    ["Base.WhiskeyFull"] = 5,
                     ["Base.Pike"] = 5,
                     ["Base.Butter"] = 5,
                     ["Base.HollyBerry"] = 5,
                     ["Base.Ginseng"] = 10,
+                },
+                fluids = {
+                    {code = "Base.Whiskey", name = "Whiskey", amount = 5},
                 },
                 result = "Moonshine.HastyHerring",
                 additionalResults = {},
@@ -95,11 +103,13 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 usedItems = {
                     ["Moonshine.ExclusiveRecipe"] = 1,
                     ["Moonshine.HastyHerring"] = 1,
-                    ["Base.WhiskeyFull"] = 5,
-                    ["Base.Pike"] = 5,
+                    ["Base.Pike"] = 10,
                     ["Base.Yeast"] = 10,
                     ["Base.Ginseng"] = 30,
                     ["Base.Violets"] = 50,
+                },
+                fluids = {
+                    {code = "Base.Whiskey", name = "Whiskey", amount = 5},
                 },
                 result = "Moonshine.DoubleHastyHerring",
                 additionalResults = {},
@@ -114,11 +124,13 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 cookingSkill = 10,
                 usedItems = {
                     ["Moonshine.ExclusiveRecipe"] = 1,
-                    ["Base.WhiskeyFull"] = 5,
                     ["Base.Sledgehammer"] = 1,
                     ["Base.WildEggs"] = 10,
                     ["Base.Salt"] = 5,
                     ["Base.Acorn"] = 5,
+                },
+                fluids = {
+                    {code = "Base.Whiskey", name = "Whiskey", amount = 5},
                 },
                 result = "Moonshine.GreedyHammer",
                 additionalResults = {},
@@ -134,11 +146,13 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 usedItems = {
                     ["Moonshine.ExclusiveRecipe"] = 1,
                     ["Moonshine.GreedyHammer"] = 1,
-                    ["Base.WhiskeyFull"] = 5,
                     ["Base.WildEggs"] = 20,
                     ["Base.Yeast"] = 10,
                     ["Base.Acorn"] = 25,
                     ["Base.Violets"] = 50,
+                },
+                fluids = {
+                    {code = "Base.Whiskey", name = "Whiskey", amount = 5},
                 },
                 result = "Moonshine.DoubleGreedyHammer",
                 additionalResults = {},
@@ -153,11 +167,13 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 cookingSkill = 10,
                 usedItems = {
                     ["Moonshine.ExclusiveRecipe"] = 1,
-                    ["Base.WhiskeyFull"] = 5,
                     ["Base.Axe"] = 1,
                     ["Base.WildEggs"] = 10,
                     ["Base.Pepper"] = 5,
                     ["Base.Acorn"] = 10,
+                },
+                fluids = {
+                    {code = "Base.Whiskey", name = "Whiskey", amount = 5},
                 },
                 result = "Moonshine.GreedyAxe",
                 additionalResults = {},
@@ -173,11 +189,13 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 usedItems = {
                     ["Moonshine.ExclusiveRecipe"] = 1,
                     ["Moonshine.GreedyAxe"] = 1,
-                    ["Base.WhiskeyFull"] = 5,
                     ["Base.WildEggs"] = 20,
                     ["Base.Yeast"] = 10,
                     ["Base.Acorn"] = 25,
                     ["Base.Violets"] = 50,
+                },
+                fluids = {
+                    {code = "Base.Whiskey", name = "Whiskey", amount = 5},
                 },
                 result = "Moonshine.DoubleGreedyAxe",
                 additionalResults = {},
@@ -192,11 +210,13 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 cookingSkill = 10,
                 usedItems = {
                     ["Moonshine.ExclusiveRecipe"] = 1,
-                    ["Base.WhiskeyFull"] = 5,
                     ["Base.Worm"] = 100,
                     ["Base.ShotgunShells"] = 20,
                     ["Base.556Bullets"] = 20,
                     ["Base.308Bullets"] = 20,
+                },
+                fluids = {
+                    {code = "Base.Whiskey", name = "Whiskey", amount = 5},
                 },
                 result = "Moonshine.StrayBullet",
                 additionalResults = {},
@@ -211,11 +231,13 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 cookingSkill = 10,
                 usedItems = {
                     ["Moonshine.ExclusiveRecipe"] = 1,
-                    ["Base.WhiskeyFull"] = 5,
-                    ["farming.RemouladeFull"] = 5,
+                    ["Base.RemouladeFull"] = 5,
                     ["Base.Butter"] = 10,
                     ["Base.OilOlive"] = 5,
                     ["Base.OilVegetable"] = 5,
+                },
+                fluids = {
+                    {code = "Base.Whiskey", name = "Whiskey", amount = 5},
                 },
                 result = "Moonshine.SlipperyFish",
                 additionalResults = {},
@@ -230,11 +252,13 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 cookingSkill = 10,
                 usedItems = {
                     ["Moonshine.ExclusiveRecipe"] = 1,
-                    ["Base.WhiskeyFull"] = 5,
-                    ["Base.Wine2"] = 5,
-                    ["farming.Bacon"] = 25,
+                    ["Base.Bacon"] = 25,
                     ["Base.Pencil"] = 10,
                     ["Base.WildGarlic2"] = 25,
+                },
+                fluids = {
+                    {code = "Base.Whiskey", name = "Whiskey", amount = 5},
+                    {code = "Base.Wine2", name = "Wine", amount = 5},
                 },
                 result = "Moonshine.SolidAdventurer",
                 additionalResults = {},
@@ -249,11 +273,13 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 cookingSkill = 10,
                 usedItems = {
                     ["Moonshine.ExclusiveRecipe"] = 1,
-                    ["Base.Wine2"] = 5,
-                    ["Base.Wine"] = 5,
                     ["Base.GrapeLeaves"] = 25,
-                    ["Base.Milk"] = 5,
                     ["Base.Pickles"] = 10,
+                },
+                fluids = {
+                    {code = "Base.Wine2", name = "Wine", amount = 5},
+                    {code = "Base.Wine", name = "Wine", amount = 5},
+                    {code = "Base.Milk", name = "CowMilk", amount = 5},
                 },
                 result = "Moonshine.SlenderDoe",
                 additionalResults = {},
@@ -268,316 +294,19 @@ function MoonshineRecipesServer.GetDefaultRecipes()
                 cookingSkill = 10,
                 usedItems = {
                     ["Moonshine.ExclusiveRecipe"] = 1,
-                    ["Base.WhiskeyFull"] = 5,
                     ["Base.Sugar"] = 10,
-                    ["Base.Cigarettes"] = 1000,
+                    ["Base.CigaretteSingle"] = 100,
                     ["Base.Coffee2"] = 5,
-                    ["Base.LemonGrass"] = 25
+                    ["Base.LemonGrass"] = 25,
+                },
+                fluids = {
+                    {code = "Base.Whiskey", name = "Whiskey", amount = 5},
                 },
                 result = "Moonshine.NicotineOverdose",
                 additionalResults = {},
             },
         },
     }
-
-    local pzversion = string.sub(getCore():getVersionNumber(), 1, 2)
-    if pzversion == "42" then
-        recipes = {
-            Vanilla = {
-                {
-                    name = "MakeWhiskey",
-                    type = "Vanilla",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "Item_WhiskeyFull",
-                    time = 300,
-                    cookingSkill = 3,
-                    usedItems = {
-                        ["Base.Yeast"] = 1,
-                        ["Base.Corn"] = 1,
-                        ["Base.UnusableWood"] = 1,
-                    },
-                    fluids = {
-                        {code = "Base.Whiskey", name = "Water", amount = 1},
-                    },
-                    result = "Base.WhiskeyFull",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeWine",
-                    type = "Vanilla",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "Item_Wine2Full",
-                    time = 300,
-                    cookingSkill = 3,
-                    usedItems = {
-                        ["Base.Sugar"] = 1,
-                        ["Base.Grapes"] = 1,
-                    },
-                    fluids = {
-                        {code = "Base.BottleCrafted", name = "Water", amount = 1},
-                    },
-                    result = "Base.Wine2",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeBeer",
-                    type = "Vanilla",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "Item_BeerBottle",
-                    time = 300,
-                    cookingSkill = 3,
-                    usedItems = {
-                        ["Base.Sugar"] = 1,
-                        ["Base.Corn"] = 1,
-                    },
-                    fluids = {
-                        {code = "Base.BeerEmpty", name = "Water", amount = 1},
-                    },
-                    result = "Base.BeerBottle",
-                    additionalResults = {},
-                },
-            },
-            Exclusive = {
-                {
-                    name = "MakeHastyHerring",
-                    type = "Exclusive",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "media/textures/Item_HastyHerring.png",
-                    time = 300,
-                    cookingSkill = 10,
-                    usedItems = {
-                        ["Moonshine.ExclusiveRecipe"] = 1,
-                        ["Base.Pike"] = 5,
-                        ["Base.Butter"] = 5,
-                        ["Base.HollyBerry"] = 5,
-                        ["Base.Ginseng"] = 10,
-                    },
-                    fluids = {
-                        {code = "Base.Whiskey", name = "Whiskey", amount = 5},
-                    },
-                    result = "Moonshine.HastyHerring",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeDoubleHastyHerring",
-                    type = "Exclusive",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "media/textures/Item_DoubleHastyHerring.png",
-                    time = 300,
-                    cookingSkill = 10,
-                    usedItems = {
-                        ["Moonshine.ExclusiveRecipe"] = 1,
-                        ["Moonshine.HastyHerring"] = 1,
-                        ["Base.Pike"] = 10,
-                        ["Base.Yeast"] = 10,
-                        ["Base.Ginseng"] = 30,
-                        ["Base.Violets"] = 50,
-                    },
-                    fluids = {
-                        {code = "Base.Whiskey", name = "Whiskey", amount = 5},
-                    },
-                    result = "Moonshine.DoubleHastyHerring",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeGreedyHammer",
-                    type = "Exclusive",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "media/textures/Item_GreedyHammer.png",
-                    time = 300,
-                    cookingSkill = 10,
-                    usedItems = {
-                        ["Moonshine.ExclusiveRecipe"] = 1,
-                        ["Base.Sledgehammer"] = 1,
-                        ["Base.WildEggs"] = 10,
-                        ["Base.Salt"] = 5,
-                        ["Base.Acorn"] = 5,
-                    },
-                    fluids = {
-                        {code = "Base.Whiskey", name = "Whiskey", amount = 5},
-                    },
-                    result = "Moonshine.GreedyHammer",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeDoubleGreedyHammer",
-                    type = "Exclusive",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "media/textures/Item_DoubleGreedyHammer.png",
-                    time = 300,
-                    cookingSkill = 10,
-                    usedItems = {
-                        ["Moonshine.ExclusiveRecipe"] = 1,
-                        ["Moonshine.GreedyHammer"] = 1,
-                        ["Base.WildEggs"] = 20,
-                        ["Base.Yeast"] = 10,
-                        ["Base.Acorn"] = 25,
-                        ["Base.Violets"] = 50,
-                    },
-                    fluids = {
-                        {code = "Base.Whiskey", name = "Whiskey", amount = 5},
-                    },
-                    result = "Moonshine.DoubleGreedyHammer",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeGreedyAxe",
-                    type = "Exclusive",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "media/textures/Item_GreedyAxe.png",
-                    time = 300,
-                    cookingSkill = 10,
-                    usedItems = {
-                        ["Moonshine.ExclusiveRecipe"] = 1,
-                        ["Base.Axe"] = 1,
-                        ["Base.WildEggs"] = 10,
-                        ["Base.Pepper"] = 5,
-                        ["Base.Acorn"] = 10,
-                    },
-                    fluids = {
-                        {code = "Base.Whiskey", name = "Whiskey", amount = 5},
-                    },
-                    result = "Moonshine.GreedyAxe",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeDoubleGreedyAxe",
-                    type = "Exclusive",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "media/textures/Item_DoubleGreedyAxe.png",
-                    time = 300,
-                    cookingSkill = 10,
-                    usedItems = {
-                        ["Moonshine.ExclusiveRecipe"] = 1,
-                        ["Moonshine.GreedyAxe"] = 1,
-                        ["Base.WildEggs"] = 20,
-                        ["Base.Yeast"] = 10,
-                        ["Base.Acorn"] = 25,
-                        ["Base.Violets"] = 50,
-                    },
-                    fluids = {
-                        {code = "Base.Whiskey", name = "Whiskey", amount = 5},
-                    },
-                    result = "Moonshine.DoubleGreedyAxe",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeStrayBullet",
-                    type = "Exclusive",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "media/textures/Item_StrayBullet.png",
-                    time = 300,
-                    cookingSkill = 10,
-                    usedItems = {
-                        ["Moonshine.ExclusiveRecipe"] = 1,
-                        ["Base.Worm"] = 100,
-                        ["Base.ShotgunShells"] = 20,
-                        ["Base.556Bullets"] = 20,
-                        ["Base.308Bullets"] = 20,
-                    },
-                    fluids = {
-                        {code = "Base.Whiskey", name = "Whiskey", amount = 5},
-                    },
-                    result = "Moonshine.StrayBullet",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeSlipperyFish",
-                    type = "Exclusive",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "media/textures/Item_SlipperyFish.png",
-                    time = 300,
-                    cookingSkill = 10,
-                    usedItems = {
-                        ["Moonshine.ExclusiveRecipe"] = 1,
-                        ["Base.RemouladeFull"] = 5,
-                        ["Base.Butter"] = 10,
-                        ["Base.OilOlive"] = 5,
-                        ["Base.OilVegetable"] = 5,
-                    },
-                    fluids = {
-                        {code = "Base.Whiskey", name = "Whiskey", amount = 5},
-                    },
-                    result = "Moonshine.SlipperyFish",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeSolidAdventurer",
-                    type = "Exclusive",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "media/textures/Item_SolidAdventurer.png",
-                    time = 300,
-                    cookingSkill = 10,
-                    usedItems = {
-                        ["Moonshine.ExclusiveRecipe"] = 1,
-                        ["Base.Bacon"] = 25,
-                        ["Base.Pencil"] = 10,
-                        ["Base.WildGarlic2"] = 25,
-                    },
-                    fluids = {
-                        {code = "Base.Whiskey", name = "Whiskey", amount = 5},
-                        {code = "Base.Wine2", name = "Wine", amount = 5},
-                    },
-                    result = "Moonshine.SolidAdventurer",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeSlenderDoe",
-                    type = "Exclusive",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "media/textures/Item_SlenderDoe.png",
-                    time = 300,
-                    cookingSkill = 10,
-                    usedItems = {
-                        ["Moonshine.ExclusiveRecipe"] = 1,
-                        ["Base.GrapeLeaves"] = 25,
-                        ["Base.Pickles"] = 10,
-                    },
-                    fluids = {
-                        {code = "Base.Wine2", name = "Wine", amount = 5},
-                        {code = "Base.Wine", name = "Wine", amount = 5},
-                        {code = "Base.Milk", name = "CowMilk", amount = 5},
-                    },
-                    result = "Moonshine.SlenderDoe",
-                    additionalResults = {},
-                },
-                {
-                    name = "MakeNicotineOverdose",
-                    type = "Exclusive",
-                    disabled = false,
-                    sound = "PourWaterIntoObject",
-                    texture = "media/textures/Item_NicotineOverdose.png",
-                    time = 300,
-                    cookingSkill = 10,
-                    usedItems = {
-                        ["Moonshine.ExclusiveRecipe"] = 1,
-                        ["Base.Sugar"] = 10,
-                        ["Base.CigaretteSingle"] = 100,
-                        ["Base.Coffee2"] = 5,
-                        ["Base.LemonGrass"] = 25,
-                    },
-                    fluids = {
-                        {code = "Base.Whiskey", name = "Whiskey", amount = 5},
-                    },
-                    result = "Moonshine.NicotineOverdose",
-                    additionalResults = {},
-                },
-            },
-        }
-    end
 
     return recipes
 end
@@ -590,13 +319,9 @@ function MoonshineRecipesServer.LoadRecipes()
     end
 
     local recipeFileName = "moonshine-recipes.json"
-    local pzversion = string.sub(getCore():getVersionNumber(), 1, 2)
-    if pzversion == "42" then
-        recipeFileName = "moonshine-recipes-b42.json"
-    end
 
     local recipes = MoonshineRecipesServer.ReadFile(recipeFileName)
-    if not recipes then
+    if not recipes or SandboxVars.Moonshine.ForceRecreateRecipesFile then
         recipes = MoonshineRecipesServer.GetDefaultRecipes()
 
         MoonshineRecipesServer.WriteFile(recipeFileName, recipes)